add authorisation to edit and destroy function in post controller

+++ app/Http/Controllers/PostsController.php
@@ -90,6 +90,11 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+        // Check for correct user
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error', 'Unauthorised Page.');
+        }
+
         return view('posts.edit')->with('post', $post);
     }
 
@@ -124,6 +129,11 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+
+        // Check for correct user
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error', 'Unauthorised Page.');
+        }
         $post->delete();
 
         return redirect('/posts')->with('success', 'Post Deleted');
